+++ dse_do_dashboard/main_pages/prepare_data_page_edit.py
@@ -5,6 +5,7 @@ from typing import List, Optional, Dict
 
 import pandas as pd
 import dash
+from dash.dash_table.Format import Format
 from dash.exceptions import PreventUpdate
 
 from dse_do_dashboard.main_pages.main_page import MainPage
@@ -115,6 +116,7 @@ class PrepareDataPageEdit(PrepareDataPage):
             ],
             fixed_rows={'headers': True},
             editable=editable,
+            virtualization=True,
             # fixed_columns={'headers': False, 'data': 0}, # Does NOT create a horizontal scroll bar
             filter_action="native",
             sort_action="native",
@@ -127,7 +129,8 @@ class PrepareDataPageEdit(PrepareDataPage):
                 'font_size': '12px',
                 'textAlign': 'left'},
             style_table={
-                'maxHeight': '400px',
+                'maxHeight': '600px',
+                # 'height': '800',
                 'overflowY': 'scroll'
             },
             style_header={