+++ CHANGELOG.md
@@ -4,10 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.1.2.1b0]
+## [Unreleased]## [0.1.2.2b3]
 ### Changed
+- setup.py avoids import of dse_do_utils to get __version__
+- Removed unused import in dash_common_utils causing failure in PyCharm debug mode
 
-## [0.1.2.1] - 2002-04-28
+## [0.1.2.2b2] - 2022-11-15
+### Changed
+- DoDashApp.read_scenario_tables_from_db_cached only reads tables in the schema, skips others without warning.
+- Enabled virtualization in DashTables (Prepare data and Explore Solution)
+- Remove restriction on SQLAlchemy version < 1.4
+
+## [0.1.2.2b1] - 2022-10-17
+### Changed
+- DashApp.run_server(self, **kwargs) with arguments for dash app.run_server()
+- Enabled write output to DB after run model.
+### Removed
+- DashApp.set_run_server_kwargs()
+
+## [0.1.2.2b0] - 2022-09-22
+### Changed
+- Stop Server button now uses os.kill() instead of deprecated Werkzeug callback
+
+## [0.1.2.1] - 2022-04-28
 ### Changed
 ### Added
 - Customize the 'brand' name of the dashboard by overriding DashApp.get_navbar_brand_children